Fixed bucketing_utils.cpp warnings.

+++ src/bucketing_utils.cpp
@@ -9,7 +9,7 @@
 #include "bucketing_constants.h"
 
 std::pair<uint64_t , uint64_t > find_percentile_position(int percentile, std::vector<uint64_t > buckets, uint64_t  buckets_total_items) {
-    uint64_t percentile_position = (buckets_total_items - 1) * ((double) percentile / (double) 100);
+    uint64_t percentile_position = static_cast<uint64_t>((buckets_total_items - 1) * ((double) percentile / (double) 100));
 
     uint64_t item_count = 0;
 
@@ -45,7 +45,7 @@ std::pair<uint64_t , uint64_t > find_percentile_position_in_subbucket(uint64_t p
 
     uint64_t  item_count = 0;
 
-    int64_t bucket_index = buckets.size() - 1;
+    size_t bucket_index = buckets.size() - 1;
     bool bucket_found = false;
     if (base_bucket >= BUCKET_COUNT / 2) {
         for (bucket_index = buckets.size() - 1; bucket_index >= 0; bucket_index--) {
